Combined a bunch of initializers for more controlled initialization order. Fixed casing for serial protocols tab tooltip id in registration.

+++ src/main/java/li/cil/tis3d/client/renderer/tileentity/TileEntitySpecialRendererCasing.java
@@ -204,7 +204,7 @@ public final class TileEntitySpecialRendererCasing extends BlockEntityRenderer<T
 
         final int brightness = getWorld().getLightmapIndex(
             casing.getPosition().offset(Face.toEnumFacing(face)), 0);
-        GLX.glMultiTexCoord2f(GLX.GL_TEXTURE1, brightness % 65536, brightness / 65536);
+        GLX.glMultiTexCoord2f(GLX.GL_TEXTURE1, brightness % 65536, (float) (brightness / 65536));
 
         try {
             module.render(renderManager, partialTicks);
